revert persistence extraction for merger
